Refactor `fetchEmails` by extracting search and parsing logic into reusable utilities
